Separate JSON utils
This allows for easy reuse of the JSON utils in the codebase, but also has the added benefit that applications can use the same JSON utils Binder use.
